requirements.txt
langchain
lark
tiktoken
openai
gradio
fastapi
uvicorn
path
azure-cognitiveservices-speech
google-api-python-client
google-search-results
wikipedia
python-dotenv
nest_asyncio
chromadb
faiss-cpu
torch
torchvision
transformers
sentence_transformers
lxml
tabulate
markdownify
